# investigate: ristretto cache with memory bounding